restructure tests
