+++ packages/react/src/dialog/popup/DialogPopup.test.tsx
@@ -234,7 +234,7 @@ describe('<Dialog.Popup />', () => {
       expect(nestedDialog).not.to.have.attribute('data-has-nested-dialogs');
     });
 
-    it('adds the `nested` and `has-nested-dialogs` style hooks if a dialog has a parent alert dialog', async () => {
+    it.only('adds the `nested` and `has-nested-dialogs` style hooks if a dialog has a parent alert dialog', async () => {
       await render(
         <AlertDialog.Root open>
           <AlertDialog.Portal>
@@ -263,35 +263,5 @@ describe('<Dialog.Popup />', () => {
       expect(parentDialog).to.have.attribute('data-has-nested-dialogs');
       expect(nestedDialog).not.to.have.attribute('data-has-nested-dialogs');
     });
-  });
-
-  it('adds the `nested` and `has-nested-dialogs` style hooks on an alert dialog if has a parent dialog', async () => {
-    await render(
-      <Dialog.Root open>
-        <Dialog.Portal>
-          <Dialog.Popup data-testid="parent-dialog" />
-          <AlertDialog.Root open>
-            <AlertDialog.Portal>
-              <AlertDialog.Popup data-testid="nested-dialog">
-                <AlertDialog.Root>
-                  <AlertDialog.Portal>
-                    <AlertDialog.Popup />
-                  </AlertDialog.Portal>
-                </AlertDialog.Root>
-              </AlertDialog.Popup>
-            </AlertDialog.Portal>
-          </AlertDialog.Root>
-        </Dialog.Portal>
-      </Dialog.Root>,
-    );
-
-    const parentDialog = screen.getByTestId('parent-dialog');
-    const nestedDialog = screen.getByTestId('nested-dialog');
-
-    expect(parentDialog).not.to.have.attribute('data-nested');
-    expect(nestedDialog).to.have.attribute('data-nested');
-
-    expect(parentDialog).to.have.attribute('data-has-nested-dialogs');
-    expect(nestedDialog).not.to.have.attribute('data-has-nested-dialogs');
   });
 });
